+++ client.d.ts
@@ -1,0 +1,5 @@
+declare module '*.svg?component' {
+  import { DefineComponent, SVGAttributes } from 'vue'
+  const component: DefineComponent<SVGAttributes>
+  export default component
+}