+++ example/env.d.ts
@@ -1,0 +1,7 @@
+/// <reference types="vite/client" />
+
+declare module '*.svg?component' {
+  import { DefineComponent, SVGAttributes } from 'vue'
+  const component: DefineComponent<SVGAttributes>
+  export default component
+}