+++ CHANGELOG.md
@@ -1,1 +1,9 @@
 # CHANGE LOG
+
+## 1.1.0 (2022-12-20)
+
+**Features**
+
+1. Support `assetsDirName` to Limit the svg icon in a folder.
+2. Add declaration file for Vue app. [client.d.ts]('./client.d.ts')
+3. Change `defaultExport` to `url`